Move sqlite foreign key pragma to __init__
