+++ asreview/__init__.py
@@ -12,6 +12,9 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from sqlalchemy import event
+from sqlalchemy.engine import Engine
+
 from asreview.data.loader import load_dataset
 from asreview.extensions import extensions
 from asreview.extensions import get_extension
@@ -51,3 +54,20 @@ __all__ = [
     "ProjectError",
     "ProjectNotFoundError",
 ]
+
+
+@event.listens_for(Engine, "connect")
+def set_sqlite_pragma(dbapi_connection, connection_record):
+    """Configure SQLite to use foreign keys.
+
+    By adding this function, everytime a connection is made to the sqlite engine, we
+    make sure that foreign keys are configured (by default sqlite allows foreign keys,
+    but ignores them and it's only SQLAlchemy that takes care of foreign keys). The only
+    downside of enabling foreign keys on the database level is that we will run into
+    problems if we have mutually dependent foreign keys.
+
+    See also: https://docs.sqlalchemy.org/en/20/dialects/sqlite.html#sqlite-foreign-keys
+    """
+    cursor = dbapi_connection.cursor()
+    cursor.execute("PRAGMA foreign_keys=ON")
+    cursor.close()
